+++ integreat_cms/cms/forms/regions/region_form.py
@@ -2,6 +2,7 @@ import logging
 
 from copy import deepcopy
 
+from zoneinfo import available_timezones
 from django import forms
 from django.conf import settings
 from django.utils.translation import override, ugettext_lazy as _
@@ -15,7 +16,46 @@ from ...utils.translation_utils import ugettext_many_lazy as __
 from ..icon_widget import IconWidget
 from ..custom_model_form import CustomModelForm
 
+
 logger = logging.getLogger(__name__)
+
+
+def get_timezone_choices():
+    """
+    This method generates the options for the second timezone dropdown
+
+    :return: A list of all available timezones
+    :rtype: list
+    """
+    timezones = list(available_timezones())
+    timezones.sort()
+    return [("", "---------")] + [
+        (tz, tz.split("/", 1)[1].replace("_", " ")) for tz in timezones if "/" in tz
+    ]
+
+
+def get_timezone_area_choices():
+    """
+    This method generates the options for the first timezone dropdown.
+    It displays the general area of a country or city. Often the continent.
+
+    :return: A list of the general areas of the timezones
+    :rtype: list
+
+    """
+    timezone_regions = list(
+        {
+            tz.split("/")[0]
+            for tz in available_timezones()
+            if "/" in tz and "Etc" not in tz and "SystemV" not in tz
+        }
+    )
+    timezone_regions.sort()
+    return (
+        [("", "---------")]
+        + [(tzr, tzr) for tzr in timezone_regions]
+        + [("Etc", _("Other timezones"))]
+    )
 
 
 class RegionForm(CustomModelForm):
@@ -26,6 +66,12 @@ class RegionForm(CustomModelForm):
     duplicated_region = forms.ModelChoiceField(
         queryset=Region.objects.all(),
         empty_label=_("Do no import initial content"),
+        required=False,
+    )
+
+    timezone_area = forms.ChoiceField(
+        choices=get_timezone_area_choices,
+        label=_("Timezone area"),
         required=False,
     )
 
@@ -62,9 +108,11 @@ class RegionForm(CustomModelForm):
             "short_urls_enabled",
             "custom_prefix",
             "tunews_enabled",
+            "timezone",
         ]
         #: The widgets which are used in this form
         widgets = {
+            "timezone": forms.Select(choices=get_timezone_choices()),
             "icon": IconWidget(),
             "offers": forms.CheckboxSelectMultiple(),
         }
@@ -80,7 +128,8 @@ class RegionForm(CustomModelForm):
         :type \**kwargs: dict
         """
         super().__init__(*args, **kwargs)
-
+        if self.instance and "/" in self.instance.timezone:
+            self.fields["timezone_area"].initial = self.instance.timezone.split("/")[0]
         self.fields["slug"].required = False
 
     def save(self, commit=True):