+++ integreat_cms/core/middleware/__init__.py
@@ -3,3 +3,4 @@ This package contains custom middlewares, see :doc:`topics/http/middleware` and 
 """
 from .access_control_middleware import AccessControlMiddleware
 from .region_middleware import RegionMiddleware
+from .timezone_middleware import TimezoneMiddleware