+++ integreat_cms/static/src/js/regions/conditional-fields.ts
@@ -28,4 +28,29 @@ window.addEventListener('load', () => {
             }
         });
     }
+    // add conditional logic to display options within the second timezone dropdown
+    const timezoneAreaDropdown = document.getElementById("id_timezone_area") as HTMLSelectElement;
+    const timezoneDropdown = document.getElementById("id_timezone") as HTMLSelectElement;
+        
+    if(timezoneDropdown && timezoneAreaDropdown) {
+        updateTimezoneAreaDropdown();
+        timezoneAreaDropdown.addEventListener('change', updateTimezoneAreaDropdown)
+    }
 });
+
+function updateTimezoneAreaDropdown(event?: Event) {
+    const timezoneAreaDropdown = document.getElementById("id_timezone_area") as HTMLSelectElement;
+    const timezoneDropdown = document.getElementById("id_timezone") as HTMLSelectElement;
+    // Reset timezone value when timezone region is changed
+    if (event !== undefined) {
+        timezoneDropdown.value = "";
+    }
+    // Hide all sub-timzones of other areas
+    timezoneDropdown
+        .querySelectorAll('option:not([value*="' + timezoneAreaDropdown.value + '"],[value=""])')
+        .forEach((option) => option.classList.add("hidden"));
+    // Show all sub-timezones of the selected area
+    timezoneDropdown
+        .querySelectorAll('option[value*="' + timezoneAreaDropdown.value + '"]')
+        .forEach((option) => option.classList.remove("hidden"));
+}