+++ tests/cms/test_duplicate_regions.py
@@ -29,6 +29,7 @@ def test_duplicate_regions(load_test_data, admin_client):
             "longitude": 1,
             "latitude": 1,
             "duplicated_region": 1,
+            "timezone": "Europe/Berlin",
         },
     )
     print(response.headers)